feat: add python alpine dockerfile and restructure folders
