+++ src/auth/auth.example.yml
