+++ src/auth/gateio_auth.py
